# frontend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/admin/products": {
      "post": {
        "operationId": "ProductsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Admin - Products"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "ProductsController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin - Products"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/admin/products/{id}": {
      "get": {
        "operationId": "ProductsController_findById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin - Products"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "operationId": "ProductsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin - Products"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "ProductsController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin - Products"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/admin/products/{id}/variations": {
      "post": {
        "operationId": "ProductsController_addVariation",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddVariationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Admin - Products"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/admin/products/variations/{variationId}": {
      "put": {
        "operationId": "ProductsController_updateVariation",
        "parameters": [
          {
            "name": "variationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateVariationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin - Products"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "ProductsController_deleteVariation",
        "parameters": [
          {
            "name": "variationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin - Products"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/admin/products/{id}/publish": {
      "post": {
        "operationId": "ProductsController_publish",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Admin - Products"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/admin/products/{id}/unpublish": {
      "post": {
        "operationId": "ProductsController_unpublish",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Admin - Products"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/admin/products/{id}/images": {
      "post": {
        "operationId": "ProductsController_addImage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddImageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Admin - Products"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/admin/products/images/{imageId}": {
      "delete": {
        "operationId": "ProductsController_deleteImage",
        "parameters": [
          {
            "name": "imageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin - Products"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/products/{id}": {
      "get": {
        "operationId": "PublicProductsController_findById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Public - Products"
        ]
      }
    },
...
    "/cart": {
      "get": {
        "operationId": "CartsController_getCart",
        "summary": "カート取得",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CartEntity"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "cart"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "CartsController_addToCart",
        "summary": "カート追加",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddToCartDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartEntity"
                }
              }
            }
          }
        },
        "tags": [
          "cart"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/cart/{id}": {
      "put": {
        "operationId": "CartsController_updateItem",
        "summary": "カートアイテム数量変更",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCartItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "cart"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "CartsController_removeItem",
        "summary": "カートアイテム削除",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "cart"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/payments/checkout": {
      "post": {
        "operationId": "PaymentsController_createCheckoutSession",
        "summary": "Stripe Checkout Sessionを作成してURLを返す",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCheckoutSessionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "url": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "payments"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/payments/webhook": {
      "post": {
        "operationId": "PaymentsController_handleWebhook",
        "summary": "Stripe Webhookエンドポイント（署名検証済み）",
        "parameters": [
          {
            "name": "stripe-signature",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "payments"
        ]
      }
    }
  },
  "info": {
    "title": "Todo API",
    "description": "NestJS + Swagger 学習用 Todo API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateProductDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateProductDto": {
        "type": "object",
        "properties": {}
      },
      "AddVariationDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateVariationDto": {
        "type": "object",
        "properties": {}
      },
      "AddImageDto": {
        "type": "object",
        "properties": {}
      },
...
      "ProductVariationEntity": {
        "type": "object",
        "properties": {}
      },
      "CartEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "sessionId": {
            "type": "string"
          },
          "variationId": {
            "type": "number"
          },
          "variation": {
            "$ref": "#/components/schemas/ProductVariationEntity"
          },
          "quantity": {
            "type": "number"
          },
          "reservedAt": {
            "format": "date-time",
            "type": "string"
          },
          "expiresAt": {
            "format": "date-time",
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "reserved",
              "purchased",
              "expired"
            ]
          }
        },
        "required": [
          "id",
          "sessionId",
          "variationId",
          "variation",
          "quantity",
          "reservedAt",
          "expiresAt",
          "status"
        ]
      },
      "AddToCartDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateCartItemDto": {
        "type": "object",
        "properties": {}
      },
      "CreateCheckoutSessionDto": {
        "type": "object",
        "properties": {
          "zip": {
            "type": "string",
            "example": "123-4567"
          },
          "prefecture": {
            "type": "string",
            "example": "東京都"
          },
          "city": {
            "type": "string",
            "example": "渋谷区"
          },
          "address1": {
            "type": "string",
            "example": "渋谷1-1-1"
          },
          "address2": {
            "type": "string",
            "example": "マンション101"
          }
        },
        "required": [
          "zip",
          "prefecture",
          "city",
          "address1"
        ]
      }
    }
  }
}